🚚 feat: moved components to their folders
