+++ src/App.js
@@ -1,7 +1,7 @@
-import CreateCustomer from "./CreateCustomer";
-import Customer from "./Customer";
-import AccountOperations from "./AccountOperations";
-import BalanceDisplay from "./BalanceDisplay";
+import CreateCustomer from './features/customers/CreateCustomer';
+import Customer from './features/customers/Customer';
+import AccountOperations from './features/accounts/AccountOperations';
+import BalanceDisplay from './features/accounts/BalanceDisplay';
 
 function App() {
   return (
